SokHub/AI_Assistant/requirements.txt
django>=4.2
googlemaps
polyline
folium
geopy
SpeechRecognition
pyttsx3
gTTS
openai-whisper
numpy
librosa
requests
pyaudio
